move zone in a separate file + small fixes PR review
