+++ src/CriteriaModal.css
@@ -7,7 +7,7 @@
   background-color: rgba(0, 0, 0, 0.5);
   padding: 50px;
 }
-#map-modal-top {
+#criteria-modal-top-noscroll {
   background-color: white;
   padding: 10px;
 }
